# Conflicts:
#	app/src/main/java/com/amsterdam/aflami/di/useCaseModule.kt
#	app/src/main/java/com/amsterdam/aflami/di/viewModelModule.kt
#	domain/src/main/java/com/amsterdam/domain/exceptions/AflamiException.kt
#	repository/src/main/java/com/amsterdam/repository/datasource/remote/MovieRemoteSource.kt
#	ui/src/main/java/com/amsterdam/ui/navigation/NavGraph.kt
#	ui/src/main/java/com/amsterdam/ui/navigation/Route.kt

+++ ui/src/main/java/com/amsterdam/ui/screens/movieDetails/MovieDetailsScreen.kt
@@ -327,7 +327,7 @@ fun MovieContent(
                 onSelectedListChange = movieDetailsInteractionListener::onSelectedListChange,
                 onAddToSelectedList = { listId ->
                     movieDetailsInteractionListener.onSaveMovieToList(
-                        movieId = state.movieId.toInt(),
+                        movieId = state.movieId,
                         listId = listId,
                     )
                 },
@@ -528,7 +528,7 @@ private fun SearchByActorContentPreview() {
                 override fun onClickAddToList() {}
 
                 override fun onSaveMovieToList(
-                    movieId: Int, listId: Long
+                    movieId: Long, listId: Long
                 ) {
                 }
